+++ src/ext/java/org/opentripplanner/ext/fares/impl/OrcaFareService.java
@@ -10,6 +10,7 @@ import java.util.Currency;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import javax.annotation.Nullable;
 import org.opentripplanner.ext.fares.model.FareRuleSet;
 import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.model.fare.FareMedium;
@@ -545,6 +546,16 @@ public class OrcaFareService extends DefaultFareService {
       );
       itineraryFares.addFareProduct(leg, transferFareProduct);
     }
+  }
+
+  /**
+   * In the base class only the rules for a specific feed are selected and then passed to the
+   * fare engine, however here we want to explicitly compute fares across feed boundaries.
+   */
+  @Nullable
+  @Override
+  protected Collection<FareRuleSet> fareRulesForFeed(FareType fareType, String feedId) {
+    return fareRulesPerType.get(fareType);
   }
 
   /**