+++ src/main/java/org/opentripplanner/openstreetmap/model/OSMNode.java
@@ -1,8 +1,12 @@
 package org.opentripplanner.openstreetmap.model;
 
+import java.util.Set;
 import org.locationtech.jts.geom.Coordinate;
+import org.opentripplanner.street.model.StreetTraversalPermission;
 
 public class OSMNode extends OSMWithTags {
+
+  static final Set<String> MOTOR_VEHICLE_BARRIERS = Set.of("bollard", "bar", "chain");
 
   public double lat;
   public double lon;
@@ -34,13 +38,12 @@ public class OSMNode extends OSMWithTags {
     return hasTag("crossing") && "traffic_signals".equals(getTag("crossing"));
   }
 
-  /**
-   * Checks if this node is bollard
+  /* Checks if this node is a barrier which prevents motor vehicle traffic
    *
    * @return true if it is
    */
-  public boolean isBollard() {
-    return isTag("barrier", "bollard");
+  public boolean isMotorVehicleBarrier() {
+    return isOneOfTags("barrier", MOTOR_VEHICLE_BARRIERS);
   }
 
   /**
@@ -50,13 +53,25 @@ public class OSMNode extends OSMWithTags {
    */
   public boolean isBarrier() {
     return (
-      isBollard() ||
+      isMotorVehicleBarrier() ||
       isPedestrianExplicitlyDenied() ||
       isBicycleExplicitlyDenied() ||
       isMotorcarExplicitlyDenied() ||
       isMotorVehicleExplicitlyDenied() ||
       isGeneralAccessDenied()
     );
+  }
+
+  /**
+   * Consider barrier tag in  permissions. Leave the rest for the super class.
+   */
+  @Override
+  public StreetTraversalPermission overridePermissions(StreetTraversalPermission def) {
+    StreetTraversalPermission permission = def;
+    if (isMotorVehicleBarrier()) {
+      permission = permission.remove(StreetTraversalPermission.CAR);
+    }
+    return super.overridePermissions(permission);
   }
 
   @Override