+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/mappers/RaptorRequestMapper.java
@@ -1,5 +1,8 @@
 package org.opentripplanner.routing.algorithm.raptoradapter.transit.mappers;
 
+import static java.util.function.Predicate.not;
+import static java.util.stream.Collectors.collectingAndThen;
+import static java.util.stream.Collectors.toList;
 import static org.opentripplanner.raptor.api.request.Optimization.PARALLEL;
 
 import io.micrometer.core.instrument.MeterRegistry;
@@ -7,16 +10,21 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.ZonedDateTime;
 import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
 import org.opentripplanner.framework.application.OTPFeature;
 import org.opentripplanner.raptor.api.model.RaptorAccessEgress;
 import org.opentripplanner.raptor.api.model.RaptorConstants;
 import org.opentripplanner.raptor.api.request.Optimization;
+import org.opentripplanner.raptor.api.request.PassThroughPoint;
+import org.opentripplanner.raptor.api.request.PassThroughPoints;
 import org.opentripplanner.raptor.api.request.RaptorRequest;
 import org.opentripplanner.raptor.api.request.RaptorRequestBuilder;
 import org.opentripplanner.raptor.rangeraptor.SystemErrDebugLogger;
 import org.opentripplanner.routing.algorithm.raptoradapter.router.performance.PerformanceTimersForRaptor;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripSchedule;
 import org.opentripplanner.routing.api.request.RouteRequest;
+import org.opentripplanner.transit.model.site.StopLocation;
 
 public class RaptorRequestMapper {
 
@@ -105,12 +113,25 @@ public class RaptorRequestMapper {
     if (preferences.transfer().maxAdditionalTransfers() != null) {
       searchParams.numberOfAdditionalTransfers(preferences.transfer().maxAdditionalTransfers());
     }
+
+    final Optional<PassThroughPoints> passThroughPoints = request
+      .getPassThroughPoints()
+      .stream()
+      .map(p -> {
+        final int[] stops = p.stopLocations().stream().mapToInt(StopLocation::getIndex).toArray();
+        return new PassThroughPoint(stops, p.name());
+      })
+      .collect(collectingAndThen(toList(), Optional::ofNullable))
+      .filter(not(List::isEmpty))
+      .map(PassThroughPoints::new);
+
     builder.withMultiCriteria(mcBuilder -> {
       preferences
         .transit()
         .raptor()
         .relaxGeneralizedCostAtDestination()
         .ifPresent(mcBuilder::withRelaxCostAtDestination);
+      passThroughPoints.ifPresent(pt -> mcBuilder.withPassThroughPoints(pt));
     });
 
     for (Optimization optimization : preferences.transit().raptor().optimizations()) {