+++ src/test/java/org/opentripplanner/graph_builder/module/linking/LinkingTest.java
@@ -3,12 +3,11 @@ package org.opentripplanner.graph_builder.module.linking;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.opentripplanner.graph_builder.module.FakeGraph.addExtraStops;
-import static org.opentripplanner.graph_builder.module.FakeGraph.addRegularStopGrid;
-import static org.opentripplanner.graph_builder.module.FakeGraph.buildGraphNoTransit;
-import static org.opentripplanner.graph_builder.module.FakeGraph.link;
+import static org.opentripplanner.graph_builder.module.linking.TestGraph.addExtraStops;
+import static org.opentripplanner.graph_builder.module.linking.TestGraph.addRegularStopGrid;
+import static org.opentripplanner.graph_builder.module.linking.TestGraph.link;
 
-import java.net.URISyntaxException;
+import java.io.File;
 import java.util.Comparator;
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -19,6 +18,8 @@ import org.opentripplanner.TestOtpModel;
 import org.opentripplanner.framework.geometry.GeometryUtils;
 import org.opentripplanner.framework.geometry.SphericalDistanceLibrary;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
+import org.opentripplanner.graph_builder.module.osm.OsmModule;
+import org.opentripplanner.openstreetmap.OsmProvider;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.street.model.StreetTraversalPermission;
 import org.opentripplanner.street.model._data.StreetModelForTest;
@@ -29,6 +30,9 @@ import org.opentripplanner.street.model.vertex.SplitterVertex;
 import org.opentripplanner.street.model.vertex.StreetVertex;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
+import org.opentripplanner.test.support.ResourceLoader;
+import org.opentripplanner.transit.model.framework.Deduplicator;
+import org.opentripplanner.transit.service.StopModel;
 import org.opentripplanner.transit.service.TransitModel;
 
 public class LinkingTest {
@@ -106,7 +110,7 @@ public class LinkingTest {
    * We do this by building the graphs and then comparing the stop tree caches.
    */
   @Test
-  public void testStopsLinkedIdentically() throws URISyntaxException {
+  public void testStopsLinkedIdentically() {
     // build the graph without the added stops
     TestOtpModel model = buildGraphNoTransit();
     Graph g1 = model.graph();
@@ -141,6 +145,22 @@ public class LinkingTest {
         assertEquals(v1.getLon(), v2.getLon(), 1e-10);
       }
     }
+  }
+
+  /** Build a graph in Columbus, OH with no transit */
+  public static TestOtpModel buildGraphNoTransit() {
+    var deduplicator = new Deduplicator();
+    var stopModel = new StopModel();
+    var gg = new Graph(deduplicator);
+    var transitModel = new TransitModel(stopModel, deduplicator);
+
+    File file = ResourceLoader.of(LinkingTest.class).file("columbus.osm.pbf");
+    OsmProvider provider = new OsmProvider(file, false);
+
+    OsmModule osmModule = OsmModule.of(provider, gg).build();
+
+    osmModule.buildGraph();
+    return new TestOtpModel(gg, transitModel);
   }
 
   private static List<StreetTransitStopLink> outgoingStls(final TransitStopVertex tsv) {